% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_all_packages.R
\name{install_all_packages}
\alias{install_all_packages}
\title{Install all packages}
\usage{
install_all_packages()
}
\value{

}
\description{
Run this to install all the packages commonly used.
}
\examples{
\dontrun{
install_all_packages()
}
}
